+++ TaffyEngine/uiobjects.cpp
@@ -1,8 +1,8 @@
 struct UIObject : Object {
 
-	UIObject(int x, int y, int w, int h, const char* filename, zLayer z);
-	UIObject(int x, int y, int w, int h, zLayer z, Image* spritesheet, CropInfo crop);
-	UIObject(int x, int y, int w, int h, u32 c, zLayer z);
+	UIObject(float x, float y, float w, float h, const char* filename, zLayer z);
+	UIObject(float x, float y, float w, float h, zLayer z, Image* spritesheet, CropInfo crop);
+	UIObject(float x, float y, float w, float h, u32 c, zLayer z);
 	UIObject();
 	~UIObject();
 
@@ -10,17 +10,17 @@ struct UIObject : Object {
 
 UIObject::UIObject() {}
 
-UIObject::UIObject(int x, int y, int w, int h, const char* filename, zLayer z = MIDDLE)
+UIObject::UIObject(float x, float y, float w, float h, const char* filename, zLayer z = MIDDLE)
 	:Object(x, y, w, h, filename, z) {
 	ob.cameraLinked = false;
 }
 
-UIObject::UIObject(int x, int y, int w, int h, zLayer z, Image* spritesheet, CropInfo crop)
+UIObject::UIObject(float x, float y, float w, float h, zLayer z, Image* spritesheet, CropInfo crop)
 	:Object(x, y, w, h, z, spritesheet, crop) {
 	ob.cameraLinked = false;
 }
 
-UIObject::UIObject(int x, int y, int w, int h, u32 c, zLayer z = MIDDLE)
+UIObject::UIObject(float x, float y, float w, float h, u32 c, zLayer z = MIDDLE)
 	:Object(x, y, w, h, c, z) {
 	ob.cameraLinked = false;
 }
